# docs/swagger.yaml
basePath: /
definitions:
  handler.inputUsernameAndPassword:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a weather service API documentation.
  title: Weather Service API
  version: "1.0"
paths:
  /api/city:
    get:
      description: Get the list of all cities
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties: true
            type: object
      summary: Get all cities
      tags:
      - cities
  /api/city/{name}:
    get:
      description: Get information about a city by name
      parameters:
      - description: City Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties: true
            type: object
      summary: Get information about a city
      tags:
      - cities
  /api/favourite:
    get:
      description: Get the list of all favourite cities of the user
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all favourite cities
      tags:
      - favourite
    post:
      description: Add a city to the user's favourite list
      parameters:
      - description: City Name
        in: query
        name: city
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add city to favourite
      tags:
      - favourite
  /api/weather:
    get:
      consumes:
      - application/json
      description: Get full weather information for a city and date
      parameters:
      - description: City name
        in: query
        name: city
        required: true
        type: string
      - description: Date (formatted as 2006-05-06T23:59:55Z)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties: true
            type: object
      summary: Get full weather info
      tags:
      - weather
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get a token
      parameters:
      - description: Username and Password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.inputUsernameAndPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - auth
  /auth/registration:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Username and Password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.inputUsernameAndPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"